move vistype icons to assets folder
