+++ packages/app/src/assets/AreaIcon.js
