+++ packages/app/src/assets/BarIcon.js
