+++ packages/app/src/assets/BubbleIcon.js
