+++ packages/app/src/assets/ColumnIcon.js
