+++ packages/app/src/assets/GaugeIcon.js
