+++ packages/app/src/assets/PieIcon.js
