+++ packages/app/src/assets/StackedBarIcon.js
