+++ packages/app/src/assets/StackedColumnIcon.js
