+++ packages/app/src/assets/YearOnYearIcon.js
