+++ packages/app/src/components/VisualizationTypeSelector/VisualizationTypeIcon.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import ColumnIcon from './icons/ColumnIcon';
-import StackedColumnIcon from './icons/StackedColumnIcon';
-import BarIcon from './icons/BarIcon';
-import StackedBarIcon from './icons/StackedBarIcon';
-import PieIcon from './icons/PieIcon';
-import GaugeIcon from './icons/GaugeIcon';
-import LineIcon from './icons/LineIcon';
-import AreaIcon from './icons/AreaIcon';
-import RadarIcon from './icons/RadarIcon';
-import BubbleIcon from './icons/BubbleIcon';
-import YearOnYearIcon from './icons/YearOnYearIcon';
+import ColumnIcon from '../../assets/ColumnIcon';
+import StackedColumnIcon from '../../assets/StackedColumnIcon';
+import BarIcon from '../../assets/BarIcon';
+import StackedBarIcon from '../../assets/StackedBarIcon';
+import PieIcon from '../../assets/PieIcon';
+import GaugeIcon from '../../assets/GaugeIcon';
+import LineIcon from '../../assets/LineIcon';
+import AreaIcon from '../../assets/AreaIcon';
+import RadarIcon from '../../assets/RadarIcon';
+import BubbleIcon from '../../assets/BubbleIcon';
+import YearOnYearIcon from '../../assets/YearOnYearIcon';
 import {
     COLUMN,
     STACKED_COLUMN,
